requirements.txt
google-generativeai>=0.3.0
streamlit>=1.31.0
tqdm>=4.66.0
PyPDF2>=3.0.0
python-docx==0.8.11
requests>=2.31.0
rarfile==4.0
python-magic-bin==0.4.14
